organize fixtures in fedimint-testing crate
