+++ fedimint-testing/Cargo.toml
@@ -16,11 +16,27 @@ path = "src/lib.rs"
 anyhow = "1.0.65"
 async-trait = "*"
 bitcoin = "0.29.2"
+bitcoincore-rpc = "0.16.0"
+cln-rpc = "0.1.1"
 fedimint-api  = { path = "../fedimint-api" }
 fedimint-bitcoind = { path = "../fedimint-bitcoind" }
 fedimint-wallet  = { path = "../modules/fedimint-wallet" }
+fedimint-mint = { path = "../modules/fedimint-mint" }
+fedimint-ln = { path = "../modules/fedimint-ln" }
+fedimint-server = { path = "../fedimint-server" }
+fedimint-rocksdb = { path = "../fedimint-rocksdb" }
+mint-client = { path = "../client/client-lib" }
 futures = "0.3"
+itertools = "0.10.5"
+lightning = "0.0.112"
+lightning-invoice = "0.20.0"
+ln-gateway = { path = "../gateway/ln-gateway" }
 secp256k1-zkp = { version = "0.7.0", features = [ "global-context", "bitcoin_hashes" ] }
 serde = "1.0.147"
 serde_json = "1.0.89"
+tokio = { version = "1.23.0", features = ["full"] }
+tracing ="0.1.37"
+tracing-subscriber = { version = "0.3.16", features = [ "env-filter" ] }
+url = "2.3.1"
+hbbft = { git = "https://github.com/jkitman/hbbft", branch = "upgrade-threshold-crypto-libs" }
 rand = "0.8"
