+++ fedimint-testing/src/federation/mod.rs
@@ -1,49 +1,27 @@
 use std::cell::RefCell;
-use std::collections::{BTreeMap, HashMap};
-use std::env;
-use std::iter::repeat;
-use std::net::SocketAddr;
-use std::path::PathBuf;
+use std::collections::BTreeMap;
 use std::rc::Rc;
-use std::sync::atomic::AtomicU16;
-use std::sync::atomic::Ordering;
-use std::sync::Arc;
 use std::time::Duration;
 
-use async_trait::async_trait;
-use bitcoin::hashes::{sha256, Hash};
-use bitcoin::KeyPair;
-use bitcoin::{secp256k1, Address};
-use cln_rpc::ClnRpc;
-use fake::FakeLightningTest;
+use bitcoin::hashes::Hash;
 use fedimint_api::cancellable::Cancellable;
 use fedimint_api::config::ClientConfig;
-use fedimint_api::core::{
-    ConsensusItem as PerModuleConsensusItem, PluginConsensusItem, MODULE_KEY_MINT,
-    MODULE_KEY_WALLET,
-};
-use fedimint_api::db::mem_impl::MemDatabase;
+use fedimint_api::core::{MODULE_KEY_MINT, MODULE_KEY_WALLET};
 use fedimint_api::db::Database;
-use fedimint_api::net::peers::IMuxPeerConnections;
 use fedimint_api::task::TaskGroup;
 use fedimint_api::Amount;
 use fedimint_api::OutPoint;
 use fedimint_api::PeerId;
 use fedimint_api::TieredMulti;
 use fedimint_bitcoind::BitcoindRpc;
-use fedimint_ln::LightningGateway;
 use fedimint_ln::LightningModule;
 use fedimint_mint::{Mint, MintOutput};
-use fedimint_server::config::ServerConfigParams;
-use fedimint_server::config::{connect, ServerConfig};
+use fedimint_server::config::ServerConfig;
 use fedimint_server::consensus::{ConsensusOutcome, ConsensusProposal};
 use fedimint_server::consensus::{FedimintConsensus, TransactionSubmissionError};
-use fedimint_server::multiplexed::PeerConnectionMultiplexer;
-use fedimint_server::net::connect::mock::MockNetwork;
-use fedimint_server::net::connect::{Connector, TlsTcpConnector};
+use fedimint_server::epoch::ConsensusItem;
 use fedimint_server::net::peers::PeerConnector;
 use fedimint_server::{all_decoders, consensus, EpochMessage, FedimintServer};
-use fedimint_testing::btc::{fixtures::FakeBitcoinTest, BitcoinTest};
 use fedimint_wallet::config::WalletConfig;
 use fedimint_wallet::db::UTXOKey;
 use fedimint_wallet::SpendableUTXO;
@@ -52,431 +30,12 @@ use fedimint_wallet::WalletConsensusItem;
 use futures::executor::block_on;
 use futures::future::{join_all, select_all};
 use hbbft::honey_badger::Batch;
-use itertools::Itertools;
-use lightning_invoice::Invoice;
-use ln_gateway::{
-    actor::GatewayActor,
-    client::{GatewayClientBuilder, MemDbFactory, StandardGatewayClientBuilder},
-    config::GatewayConfig,
-    rpc::GatewayRequest,
-    LnGateway,
-};
-use mint_client::{
-    api::WsFederationApi, mint::SpendableNote, Client, FederationId, GatewayClient,
-    GatewayClientConfig, UserClient, UserClientConfig,
-};
-use rand::rngs::OsRng;
-use rand::RngCore;
-use real::{RealBitcoinTest, RealLightningTest};
-use tokio::sync::Mutex;
+use mint_client::mint::SpendableNote;
 use tracing::info;
-use tracing_subscriber::EnvFilter;
-use url::Url;
 
-use crate::fixtures::utils::LnRpcAdapter;
-use crate::ConsensusItem;
-
-mod fake;
-mod real;
-mod utils;
-
-static BASE_PORT: AtomicU16 = AtomicU16::new(4000_u16);
-
-// Helper functions for easier test writing
-pub fn rng() -> OsRng {
-    OsRng
-}
-
-pub fn msats(amount: u64) -> Amount {
-    Amount::from_msat(amount)
-}
-
-pub fn sats(amount: u64) -> Amount {
-    Amount::from_sat(amount)
-}
-
-pub fn sha256(data: &[u8]) -> sha256::Hash {
-    bitcoin::hashes::sha256::Hash::hash(data)
-}
-
-pub fn secp() -> secp256k1::Secp256k1<secp256k1::All> {
-    bitcoin::secp256k1::Secp256k1::new()
-}
-
-#[non_exhaustive]
-pub struct Fixtures {
-    pub fed: FederationTest,
-    pub user: UserTest<UserClientConfig>,
-    pub bitcoin: Box<dyn BitcoinTest>,
-    pub gateway: GatewayTest,
-    pub lightning: Box<dyn LightningTest>,
-    pub task_group: TaskGroup,
-}
-
-/// Generates the fixtures for an integration test and spawns API and HBBFT consensus threads for
-/// federation nodes starting at port 4000.
-pub async fn fixtures(num_peers: u16) -> anyhow::Result<Fixtures> {
-    let mut task_group = TaskGroup::new();
-    let base_port = BASE_PORT.fetch_add(num_peers * 10, Ordering::Relaxed);
-
-    // in case we need to output logs using 'cargo test -- --nocapture'
-    if base_port == 4000 {
-        tracing_subscriber::fmt()
-            .with_env_filter(
-                EnvFilter::try_from_default_env()
-                    .unwrap_or_else(|_| EnvFilter::new("info,fedimint::consensus=warn")),
-            )
-            .init();
-    }
-    let peers = (0..num_peers as u16).map(PeerId::from).collect::<Vec<_>>();
-    let params = ServerConfigParams::gen_local(
-        &peers,
-        Amount::from_sat(1000),
-        base_port,
-        "test",
-        "127.0.0.1:18443",
-    );
-    let max_evil = hbbft::util::max_faulty(peers.len());
-
-    match env::var("FM_TEST_DISABLE_MOCKS") {
-        Ok(s) if s == "1" => {
-            info!("Testing with REAL Bitcoin and Lightning services");
-            let (server_config, client_config) =
-                distributed_config(&peers, params, max_evil, &mut task_group)
-                    .await
-                    .expect("distributed config should not be canceled");
-
-            let dir = env::var("FM_TEST_DIR").expect("Must have test dir defined for real tests");
-            let wallet_config: WalletConfig = server_config
-                .iter()
-                .last()
-                .unwrap()
-                .1
-                .get_module_config("wallet")
-                .unwrap();
-            let bitcoin_rpc = fedimint_bitcoind::bitcoincore_rpc::make_bitcoind_rpc(
-                &wallet_config.btc_rpc,
-                task_group.make_handle(),
-            )
-            .expect("Could not create bitcoinrpc");
-            let bitcoin = RealBitcoinTest::new(&wallet_config.btc_rpc);
-            let socket_gateway = PathBuf::from(dir.clone()).join("ln1/regtest/lightning-rpc");
-            let socket_other = PathBuf::from(dir.clone()).join("ln2/regtest/lightning-rpc");
-            let lightning =
-                RealLightningTest::new(socket_gateway.clone(), socket_other.clone()).await;
-            let gateway_lightning_rpc = Mutex::new(
-                ClnRpc::new(socket_gateway.clone())
-                    .await
-                    .expect("connect to ln_socket"),
-            );
-            let lightning_rpc_adapter = LnRpcAdapter::new(Box::new(gateway_lightning_rpc));
-
-            let connect_gen =
-                |cfg: &ServerConfig| TlsTcpConnector::new(cfg.tls_config()).into_dyn();
-            let fed_db = || rocks(dir.clone()).into();
-            let fed = FederationTest::new(
-                server_config,
-                &fed_db,
-                &|| bitcoin_rpc.clone(),
-                &connect_gen,
-                &mut task_group,
-            )
-            .await;
-
-            let user_db = rocks(dir.clone()).into();
-            let user_cfg = UserClientConfig(client_config.clone());
-            let user = UserTest::new(Arc::new(create_user_client(user_cfg, peers, user_db).await));
-            user.client.await_consensus_block_height(0).await?;
-
-            let gateway = GatewayTest::new(
-                lightning_rpc_adapter,
-                client_config.clone(),
-                lightning.gateway_node_pub_key,
-                base_port + (2 * num_peers) + 1,
-            )
-            .await;
-
-            Ok(Fixtures {
-                fed,
-                user,
-                bitcoin: Box::new(bitcoin),
-                gateway,
-                lightning: Box::new(lightning),
-                task_group,
-            })
-        }
-        _ => {
-            info!("Testing with FAKE Bitcoin and Lightning services");
-            let (server_config, client_config) =
-                ServerConfig::trusted_dealer_gen(&peers, &params, OsRng);
-
-            let bitcoin = FakeBitcoinTest::new();
-            let bitcoin_rpc = || bitcoin.clone().into();
-            let lightning = FakeLightningTest::new();
-            let ln_rpc_adapter = LnRpcAdapter::new(Box::new(lightning.clone()));
-            let net = MockNetwork::new();
-            let net_ref = &net;
-            let connect_gen = move |cfg: &ServerConfig| net_ref.connector(cfg.identity).into_dyn();
-
-            let fed_db = || MemDatabase::new().into();
-            let fed = FederationTest::new(
-                server_config,
-                &fed_db,
-                &bitcoin_rpc,
-                &connect_gen,
-                &mut task_group,
-            )
-            .await;
-
-            let user_db = MemDatabase::new().into();
-            let user_cfg = UserClientConfig(client_config.clone());
-            let user = UserTest::new(Arc::new(create_user_client(user_cfg, peers, user_db).await));
-            user.client.await_consensus_block_height(0).await?;
-
-            let gateway = GatewayTest::new(
-                ln_rpc_adapter,
-                client_config.clone(),
-                lightning.gateway_node_pub_key,
-                base_port + (2 * num_peers) + 1,
-            )
-            .await;
-
-            Ok(Fixtures {
-                fed,
-                user,
-                bitcoin: Box::new(bitcoin),
-                gateway,
-                lightning: Box::new(lightning),
-                task_group,
-            })
-        }
-    }
-}
-
-pub fn peers(peers: &[u16]) -> Vec<PeerId> {
-    peers
-        .iter()
-        .map(|i| PeerId::from(*i))
-        .collect::<Vec<PeerId>>()
-}
-
-/// Creates a new user client connected to the given peers
-pub async fn create_user_client(
-    config: UserClientConfig,
-    peers: Vec<PeerId>,
-    db: Database,
-) -> UserClient {
-    let api = WsFederationApi::new(
-        config
-            .0
-            .nodes
-            .iter()
-            .enumerate()
-            .filter(|(id, _)| peers.contains(&PeerId::from(*id as u16)))
-            .map(|(id, node)| (PeerId::from(id as u16), node.url.clone()))
-            .collect(),
-    )
-    .into();
-
-    UserClient::new_with_api(config, db, api, Default::default()).await
-}
-
-async fn distributed_config(
-    peers: &[PeerId],
-    params: HashMap<PeerId, ServerConfigParams>,
-    _max_evil: usize,
-    task_group: &mut TaskGroup,
-) -> Cancellable<(BTreeMap<PeerId, ServerConfig>, ClientConfig)> {
-    let configs: Cancellable<Vec<(PeerId, ServerConfig)>> = join_all(peers.iter().map(|peer| {
-        let params = params.clone();
-        let peers = peers.to_vec();
-
-        let mut task_group = task_group.clone();
-
-        async move {
-            let our_params = params[peer].clone();
-            let server_conn = connect(
-                our_params.server_dkg.clone(),
-                our_params.tls.clone(),
-                &mut task_group,
-            )
-            .await;
-            let connections = PeerConnectionMultiplexer::new(server_conn).into_dyn();
-
-            let rng = OsRng;
-            let cfg = ServerConfig::distributed_gen(
-                &connections,
-                peer,
-                &peers,
-                &our_params,
-                rng,
-                &mut task_group,
-            );
-            (*peer, cfg.await.expect("generation failed"))
-        }
-    }))
-    .await
-    .into_iter()
-    .map(|(peer_id, maybe_cancelled)| maybe_cancelled.map(|v| (peer_id, v)))
-    .collect();
-
-    let configs = configs?;
-
-    let (_, config) = configs.first().unwrap().clone();
-
-    Ok((configs.into_iter().collect(), config.to_client_config()))
-}
-
-fn rocks(dir: String) -> fedimint_rocksdb::RocksDb {
-    let db_dir = PathBuf::from(dir).join(format!("db-{}", rng().next_u64()));
-    fedimint_rocksdb::RocksDb::open(db_dir).unwrap()
-}
-
-#[async_trait]
-pub trait LightningTest {
-    /// Creates invoice from a non-gateway LN node
-    async fn invoice(&self, amount: Amount, expiry_time: Option<u64>) -> Invoice;
-
-    /// Returns the amount that the gateway LN node has sent
-    async fn amount_sent(&self) -> Amount;
-}
-
-pub struct GatewayTest {
-    pub actor: Arc<GatewayActor>,
-    pub adapter: Arc<LnRpcAdapter>,
-    pub keys: LightningGateway,
-    pub user: UserTest<GatewayClientConfig>,
-    pub client: Arc<GatewayClient>,
-}
-
-impl GatewayTest {
-    async fn new(
-        ln_client_adapter: LnRpcAdapter,
-        client_config: ClientConfig,
-        node_pub_key: secp256k1::PublicKey,
-        bind_port: u16,
-    ) -> Self {
-        let mut rng = OsRng;
-        let ctx = bitcoin::secp256k1::Secp256k1::new();
-        let kp = KeyPair::new(&ctx, &mut rng);
-
-        let keys = LightningGateway {
-            mint_pub_key: kp.x_only_public_key().0,
-            node_pub_key,
-            api: Url::parse("http://example.com")
-                .expect("Could not parse URL to generate GatewayClientConfig API endpoint"),
-        };
-
-        let bind_addr: SocketAddr = format!("127.0.0.1:{}", bind_port).parse().unwrap();
-        let announce_addr = Url::parse(format!("http://{}", bind_addr).as_str())
-            .expect("Could not parse URL to generate GatewayClientConfig API endpoint");
-        let gw_client_cfg = GatewayClientConfig {
-            client_config: client_config.clone(),
-            redeem_key: kp,
-            timelock_delta: 10,
-            api: announce_addr.clone(),
-            node_pub_key,
-        };
-
-        // Create federation client builder for the gateway
-        let client_builder: GatewayClientBuilder =
-            StandardGatewayClientBuilder::new(PathBuf::new(), MemDbFactory.into()).into();
-
-        let (sender, receiver) = tokio::sync::mpsc::channel::<GatewayRequest>(100);
-        let adapter = Arc::new(ln_client_adapter);
-        let ln_rpc = Arc::clone(&adapter);
-
-        let gw_cfg = GatewayConfig {
-            bind_address: bind_addr,
-            announce_address: announce_addr,
-            password: "abc".into(),
-            default_federation: FederationId(gw_client_cfg.client_config.federation_name.clone()),
-        };
-
-        let gateway = LnGateway::new(
-            gw_cfg,
-            ln_rpc,
-            client_builder.clone(),
-            sender,
-            receiver,
-            TaskGroup::new(),
-        )
-        .await;
-
-        let client = Arc::new(
-            client_builder
-                .build(gw_client_cfg.clone())
-                .await
-                .expect("Could not build gateway client"),
-        );
-
-        let actor = gateway
-            .register_federation(client.clone())
-            .await
-            .expect("Could not register federation");
-        // Note: We don't run the gateway in test scenarios
-
-        // Create a user test from gateway federation client
-        let user = UserTest::new(client.clone());
-
-        GatewayTest {
-            actor,
-            adapter,
-            keys,
-            user,
-            client,
-        }
-    }
-}
-
-#[derive(Clone)]
-pub struct UserTest<C> {
-    pub client: Arc<Client<C>>,
-    pub config: C,
-}
-
-impl<T: AsRef<ClientConfig> + Clone> UserTest<T> {
-    pub fn new(client: Arc<Client<T>>) -> Self {
-        let config = client.config();
-        UserTest { client, config }
-    }
-
-    /// Helper to simplify the peg_out method calls
-    pub async fn peg_out(&self, amount: u64, address: &Address) -> (Amount, OutPoint) {
-        let peg_out = self
-            .client
-            .new_peg_out_with_fees(bitcoin::Amount::from_sat(amount), address.clone())
-            .await
-            .unwrap();
-        let out_point = self.client.peg_out(peg_out.clone(), rng()).await.unwrap();
-        (peg_out.fees.amount().into(), out_point)
-    }
-
-    /// Returns the amount denominations of all coins from lowest to highest
-    pub async fn coin_amounts(&self) -> Vec<Amount> {
-        self.client
-            .coins()
-            .await
-            .iter_tiers()
-            .flat_map(|(a, c)| repeat(*a).take(c.len()))
-            .sorted()
-            .collect::<Vec<Amount>>()
-    }
-
-    /// Returns sum total of all coins
-    pub async fn total_coins(&self) -> Amount {
-        self.client.coins().await.total_amount()
-    }
-
-    pub async fn assert_total_coins(&self, amount: Amount) {
-        self.client.fetch_all_coins().await;
-        assert_eq!(self.total_coins().await, amount);
-    }
-    pub async fn assert_coin_amounts(&self, amounts: Vec<Amount>) {
-        self.client.fetch_all_coins().await;
-        assert_eq!(self.coin_amounts().await, amounts);
-    }
-}
+use crate::btc::BitcoinTest;
+use crate::user::UserTest;
+use crate::{assert_module_ci, rng};
 
 pub struct FederationTest {
     servers: Vec<Rc<RefCell<ServerTest>>>,
@@ -889,7 +448,7 @@ impl FederationTest {
         }
     }
 
-    async fn new(
+    pub async fn new(
         server_config: BTreeMap<PeerId, ServerConfig>,
         database_gen: &impl Fn() -> Database,
         bitcoin_gen: &impl Fn() -> BitcoindRpc,
@@ -960,16 +519,4 @@ impl FederationTest {
             wallet,
         }
     }
-}
-
-pub fn assert_ci<M: PluginConsensusItem>(ci: &ConsensusItem) -> &M {
-    if let ConsensusItem::Module(mci) = ci {
-        assert_module_ci(mci)
-    } else {
-        panic!("Not a module consensus item");
-    }
-}
-
-pub fn assert_module_ci<M: PluginConsensusItem>(mci: &PerModuleConsensusItem) -> &M {
-    mci.as_any().downcast_ref().unwrap()
 }
